gmail/pyproject.toml
[project]
name = "gmail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "google-api-python-client>=2.188.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "httpx>=0.28.1",
    "mcp[cli]>=1.25.0",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
